AM-391 Make Authentication Provider required for a Content Provider

+++ openiam-am/openiam-am-services/src/main/java/org/openiam/am/srvc/service/AuthProviderServiceImpl.java
@@ -197,6 +197,9 @@ public class AuthProviderServiceImpl implements AuthProviderService {
         	if(CollectionUtils.isNotEmpty(entity.getContentProviders())) {
         		throw new BasicDataServiceException(ResponseCode.LINKED_TO_ONE_OR_MORE_CONTENT_PROVIDERS);
         	}
+        	if(CollectionUtils.isNotEmpty(entity.getUriPatterns())) {
+        		throw new BasicDataServiceException(ResponseCode.LINKED_TO_ONE_OR_MORE_URI_PATTERNS);
+        	}
         	authProviderDao.delete(entity);
             resourceService.deleteResource(entity.getResource().getId());
         }
